+++ src/pages/Layout/AuthLayout/AuthLayout.tsx
@@ -1,10 +1,11 @@
+import { useIsLoggedIn } from "@hooks/useIsLoggedIn";
 import { Center, Container } from "@mantine/core";
 import type { CustomLayout } from "next";
 
 import { LayoutErrorBoundary } from "../LayoutErrorBoundary";
 
 export const AuthLayout: CustomLayout = (page) => {
-  //useIsLoggedIn();
+  useIsLoggedIn();
   return (
     <Center>
       <Container size="xs" sx={{ width: 480, paddingBottom: 16 }}>
