+++ src/pages/Layout/DashboardLayout/Header/index.tsx
@@ -1,4 +1,5 @@
 import { ColorTheme } from "./ColorTheme";
 import { Lang } from "./Lang";
+import { HeaderComp } from "./HeaderComp";
 
-export { ColorTheme, Lang };
+export { ColorTheme, Lang, HeaderComp };
