+++ src/pages/_app.page.tsx
@@ -8,8 +8,8 @@ import { createEmotionCache, MantineProvider } from "@mantine/core";
 import { AppShell, Navbar, Header } from "@mantine/core";
 //import { HeaderComp } from "@components/global/Header/HeaderComp";
 import { HeadContents } from "@components/head/HeadContents";
-import { Sidebar } from "@pages/Layout/DashboardLayout/sideBar/Sidebar";
-import { User } from "@pages/Layout/DashboardLayout/sideBar/User";
+import { Sidebar } from "@pages/Layout/DashboardLayout/sideNav/SideNav";
+import { User } from "@pages/Layout/DashboardLayout/sideNav/User";
 import { state, saveUserId, saveUserEmail, saveUserName } from "@state/state";
 import dynamic from "next/dynamic";
 import { NotificationsProvider } from "@mantine/notifications";
